ref: styles are kept separate from their parent layouts
