+++ src/base/style.ts
@@ -14,54 +14,3 @@ export interface DocumentStyle extends FontStyle {
    */
   background?: Drawable;
 }
-
-export interface ImageStyle {
-  /**
-   * width of image
-   */
-  width: number;
-  /**
-   * Height of image
-   */
-  height: number;
-}
-
-export interface LineBreakStyle {
-  /**
-   * height of space to leave
-   */
-  height: number;
-}
-
-export interface BlockStyle {
-  display: "block";
-  /**
-   * Margins for this layout if any
-   */
-  margin?: CSSMargins;
-  /**
-   * Background for this layout if any
-   */
-  background?: Drawable;
-}
-
-export interface FlexStyle {
-  display: "flex";
-  /**
-   * Margins for this layout if any
-   */
-  margin?: CSSMargins;
-  /**
-   * Background for this layout if any
-   */
-  background?: Drawable;
-}
-
-/**
- * How should each item use space around it
- */
-export type FlexFloat = "left" | "right" | "none";
-/**
- * What width should an item be
- */
-export type FlexWidth = "flex" | "auto" | number;
