+++ src/layouts/block.ts
@@ -1,6 +1,19 @@
 import sumBy from "lodash/sumBy";
-import { BlockStyle, BoundingBox, Context, Element } from "../base";
+import { BoundingBox, Context, Drawable, Element } from "../base";
 import { removeMargins } from "../base/boundingBox";
+import { CSSMargins } from "../utils/margin";
+
+export interface BlockStyle {
+  display: "block";
+  /**
+   * margins for this layout if any
+   */
+  margin?: CSSMargins;
+  /**
+   * background for this layout if any
+   */
+  background?: Drawable;
+}
 
 /**
  * This arranges all it's element on a vertical line using the width
